+++ main.py
@@ -2,31 +2,37 @@ import pygame
 import sys
 import modules.input
 from modules.entities import PhysicsEntity
-from modules.utils import load_image
+from modules.utils import load_image, load_images
+from modules.tilemap import Tilemap
 def distance(A, B): return (sum(((B[i] - A[i])**2 for i in range(2))))**0.5
 
 class Game(modules.input.Input):
     def __init__(self):
         modules.input.Input.__init__(self)
-        self.width = 640
-        self.height = 480
-        self.velocity = 5 
-        self.jumpcount = 10
-        self.isJumping = False
-        self.negative = 1 
+        self.width = 1280
+        self.height = 960
+
         pygame.init()
         pygame.display.set_caption("coffeboy e watergirl")
 
         self.screen = pygame.display.set_mode((self.width, self.height))
-        self.las_pressed_pos = tuple()
+        #gera uma imagem. Pra aumentar o tamanho das coisas na tela,
+        #renderizamos nela e depois escalamos pra screen 
+        self.display = pygame.Surface((320, 240))
+
         self.player_x, self.player_y = self.width // 2, self.height // 2
         self.movement = [False, False]
 
-        #loading everything's sprites
+        #loads the images as a list of assets containing every variant of that type
         self.assets = {
+            'decor': load_images('tiles/decor'),
+            'grass': load_images('tiles/grass'),
+            'large_decor': load_images('tiles/large_decor'),
+            'stone': load_images('tiles/stone'),
             'player': load_image('entities/player/idle/00.png')}
-
+        print(self.assets)
         self.player = PhysicsEntity(self, 'player', (50, 50), (8, 15))
+        self.tilemap = Tilemap(self, tile_size=16)
 
 
 
@@ -34,10 +40,11 @@ class Game(modules.input.Input):
         clock = pygame.time.Clock()
 
         while True:
-            self.screen.fill((200,200,255))
+            self.display.fill((200,200,255))
 
+            self.tilemap.render(self.display)
             self.player.update((self.movement[1] - self.movement[0], 0))
-            self.player.render(self.screen)
+            self.player.render(self.display)
 
             for event in pygame.event.get():
                 if event.type == pygame.QUIT:
@@ -54,10 +61,11 @@ class Game(modules.input.Input):
                     if event.key == pygame.K_RIGHT:
                         self.movement[1] = False
 
-            pygame.draw.rect(self.screen, (255,0,0), (self.player_x, self.player_y, 10, 16))
+            pygame.draw.rect(self.display, (255,0,0), (self.player_x, self.player_y, 10, 16))
 
-            #isto aqui é o que desenha na tela as alterações que fazemos a cada
-            # iter do loop
+            # isto aqui é o que escala o display pra screen (A tela de vdd)
+            # e escreve na tela as alteracoes que fizemos, a cada iter
+            self.screen.blit(pygame.transform.scale(self.display, self.screen.get_size()), (0,0))
             pygame.display.update()
             clock.tick(60)